Refactor location of bytecode circuit
