+++ zkevm-circuits/src/bytecode_circuit/dev.rs
@@ -1,6 +1,5 @@
-use super::bytecode_unroller::{
-    unroll, BytecodeCircuit, BytecodeCircuitConfig, BytecodeCircuitConfigArgs, UnrolledBytecode,
-};
+use super::bytecode_unroller::{unroll, UnrolledBytecode};
+use super::circuit::{BytecodeCircuit, BytecodeCircuitConfig, BytecodeCircuitConfigArgs};
 use crate::table::{BytecodeTable, KeccakTable};
 use crate::util::{Challenges, SubCircuit, SubCircuitConfig};
 use eth_types::Field;
